unifocus/nlp-service/requirements.txt
# Python依赖包
# FastAPI 和相关依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# NLP 工具
spacy==3.7.2
sentence-transformers==2.2.2
jieba==0.42.1

# OCR 引擎
paddleocr==2.7.0
paddlepaddle==2.5.2
pytesseract==0.3.10
Pillow==10.1.0

# HTTP 客户端
httpx==0.25.1
requests==2.31.0

# 数据处理
numpy==1.24.3
pandas==2.1.3

# HTML解析
beautifulsoup4==4.12.2
lxml==4.9.3

# PDF处理
pdfminer.six==20221105

# Word文档处理
python-docx==1.1.0

# 日志
loguru==0.7.2

# 其他工具
python-dotenv==1.0.0
pyyaml==6.0.1
